policy_gui/requirements.txt
actionlib==1.11.13
adium-theme-ubuntu==0.3.4
angles==1.9.11
attrs==15.2.0
autobahn==0.10.3
backports.functools-lru-cache==1.5
base-local-planner==1.14.4
bondpy==1.8.3
bzr==2.7.0
camera-calibration==1.12.23
camera-calibration-parsers==1.11.13
catkin==0.7.14
catkin-pkg==0.4.9
catkin-pkg-modules==0.4.9
chardet==2.3.0
configobj==5.0.6
controller-manager==0.13.3
controller-manager-msgs==0.13.3
cryptography==1.2.3
cv-bridge==1.12.8
cycler==0.10.0
defusedxml==0.4.1
diagnostic-analysis==1.9.3
diagnostic-common-diagnostics==1.9.3
diagnostic-updater==1.9.3
docutils==0.12
dynamic-reconfigure==1.5.50
ecdsa==0.13
empy==3.3.2
enum34==1.1.2
futures==3.0.5
gazebo-plugins==2.5.17
gazebo-ros==2.5.17
gencpp==0.6.0
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.11
gennodejs==2.0.1
genpy==0.6.7
httplib2==0.9.1
idna==2.0
image-geometry==1.12.8
interactive-markers==1.11.4
ipaddress==1.0.16
keyboard==0.13.2
keyring==7.3
kiwisolver==1.0.1
laser-geometry==1.6.4
launchpadlib==1.10.3
lazr.restfulclient==0.13.4
lazr.uri==1.0.3
lz4==0.7.0
matplotlib==2.2.3
mercurial==3.7.3
message-filters==1.12.14
mpi4py==1.3.1
msgpack-python==0.4.6
netifaces==0.10.4
nose==1.3.7
numpy==1.15.2
oauth==1.0.1
PAM==0.4.2
pandas==0.23.4
paramiko==1.16.0
Pillow==3.1.2
pluginlib==1.11.3
psutil==3.4.2
pyasn1==0.1.9
pyasn1-modules==0.0.7
pycrypto==2.6.1
pydot==1.0.29
Pygments==2.1
pygobject==3.20.0
pygpgme==0.3
pynput==1.4
PyOpenGL==3.0.2
pyOpenSSL==0.15.1
pyparsing==2.2.0
pyserial==3.0.1
Pyste==0.9.10
python-dateutil==2.7.3
python-qt-binding==0.3.4
python-snappy==0.5
python-xlib==0.25
pytz==2018.5
PyYAML==3.11
qt-dotgraph==0.3.11
qt-gui==0.3.11
qt-gui-cpp==0.3.11
qt-gui-py-common==0.3.11
readchar==2.0.1
resource-retriever==1.12.4
roman==2.0.0
rosbag==1.12.14
rosboost-cfg==1.14.4
rosclean==1.14.4
roscreate==1.14.4
rosdep==0.12.2
rosdistro==0.6.9
rosdistro-modules==0.6.9
rosgraph==1.12.14
roslaunch==1.12.14
roslib==1.14.4
roslint==0.11.0
roslz4==1.12.14
rosmake==1.14.4
rosmaster==1.12.14
rosmsg==1.12.14
rosnode==1.12.14
rosparam==1.12.14
rospkg==1.1.7
rospkg-modules==1.1.7
rospy==1.12.14
rosservice==1.12.14
rostest==1.12.14
rostopic==1.12.14
rosunit==1.14.4
roswtf==1.12.14
rqt-action==0.4.9
rqt-bag==0.4.12
rqt-bag-plugins==0.4.12
rqt-console==0.4.8
rqt-dep==0.4.9
rqt-graph==0.4.9
rqt-gui==0.5.0
rqt-gui-py==0.5.0
rqt-image-view==0.4.13
rqt-launch==0.4.8
rqt-logger-level==0.4.8
rqt-moveit==0.5.7
rqt-msg==0.4.8
rqt-nav-view==0.5.7
rqt-plot==0.4.8
rqt-pose-view==0.5.8
rqt-publisher==0.4.8
rqt-py-common==0.5.0
rqt-py-console==0.4.8
rqt-reconfigure==0.4.10
rqt-robot-dashboard==0.5.7
rqt-robot-monitor==0.5.8
rqt-robot-steering==0.5.9
rqt-runtime-monitor==0.5.7
rqt-rviz==0.5.10
rqt-service-caller==0.4.8
rqt-shell==0.4.9
rqt-srv==0.4.8
rqt-tf-tree==0.5.8
rqt-top==0.4.8
rqt-topic==0.4.10
rqt-web==0.4.8
rviz==1.12.16
scipy==1.1.0
SecretStorage==2.1.3
sensor-msgs==1.12.6
service-identity==16.0.0
simplejson==3.8.1
six==1.11.0
smach==2.0.1
smach-ros==2.0.1
smclib==1.8.3
subprocess32==3.5.2
tf==1.11.9
tf-conversions==1.11.9
tf2-geometry-msgs==0.5.18
tf2-kdl==0.5.18
tf2-py==0.5.18
tf2-ros==0.5.18
topic-tools==1.12.14
trollius==2.0.1
Twisted==16.0.0
txaio==1.0.0
unity-lens-photos==1.0
vcstools==0.1.40
wadllib==1.3.2
wstool==0.1.17
wxPython==3.0.2.0
wxPython-common==3.0.2.0
xacro==1.11.3
zope.interface==4.1.3
